requirements.txt
django-model-utils
Werkzeug
django-extensions
ipython
bpython
ptpython
django-reversion
django-reversion-extras
# cleanerversion
#-e git://github.com/luzfcb/django-versioning.git#egg=versioning
django-debug-toolbar
# -e git://github.com/treyhunner/django-simple-history.git#egg=django-simple-history
# django-simple-history
django-crispy-forms
django-bootstrap-pagination
autopep8
